add atfork module and use it to reset hub in child on fork
this is particularly important for epoll and kqueue hubs. if they wind
up being shared by parent and child processes after a fork, the two
processes's conflict with each other because they wind up using the same
file descriptors to refer to different resources. this change gives the
child process a fresh hub of its own after a fork.