wsgi: Provide python logging compatibility
Provide a compatibility layer in eventlet.wsgi that allows you
to hand a python logger as the log parameter instead of a
filehandle.

If an object that looks like a python logger is passed, we use
it as expected. If not, we assume this is a fileobject that
implements the write() call, and wrap it in a compatiblity layer
that lets us user log() and debug() calls on it, including
optional *args.

https://github.com/eventlet/eventlet/pull/75