Allow more gracefull shutdown of wsgi server
Currently, if the listening socket is closed the
wsgi.server will go into an infinite loop.

Add an 'is_accepting' flag which can be set to False
to tell the server to exit. This allows, for example,
a signal handler to initiate a gracefull shutdown
(complete existing requests then exit).