greenpool: Make sure that imap/starmap set the correct pool
Calling a GreenPool.imap() followed by a GreenPool.waitall() does not in fact
wait on the threads that were spawned, this is because we create a
completely new GreanPool instance for the resulting GreenPile result.

We really want to be spawning threads limited by the GreenPool instance
we call the imap() on, so we should be using it for the resulting
GreenPile.