Fix GreenFileIO.readall() for regular file
GreenFileIO.readall() calls _trampoline(self, read=True). Problem: on
regular file, epoll_ctl() fails with EPERM on Linux, because regular
files are not supported.

This change skips the call to _trampoline() on regular files. Add
also various unit tests on the various ways to read the whole content
of a file.