tests: half-hearted fix sporadic zmq_test failures on Travis
Real fix must work without any sleep(), even 0 delay.
With all these sync events the code makes perfect sense and execution is properly organized.
So I am 99% sure there is a problem in green/zmq.
Hint for the person willing to take on this in future:
send/recv seems not to switch to zmq trampolining