Introduce dagpool.PropagateError. Wrap uncaught exceptions with it.
This changeset introduces a mechanism for propagating uncaught exceptions
through the graph of dependencies. If a DAGPool greenthread terminates with an
exception, that exception is wrapped in PropagateError. Attempting to retrieve
the value for that greenthread's key raises that PropagateError instance. This
may, in turn, cause a consuming greenthread to terminate with that
PropagateError...

In this way an upstream failure can be propagated through downstream
dependents in the entire rest of the DAG.

Introduce wait_each_success() to retrieve all successful results. Since
waitall(), wait() and wait_each() can raise PropagateError if any greenthread
terminated with an exception, this provides a way to harvest everything that
did succeed.

Similarly, introduce wait_each_exception(). This yields (rather than raising)
any captured PropagateError instances. This allows enumerating ALL failures,
rather than (an arbitrary) first one.

Make wait()'s keys argument optional; if omitted, it behaves like waitall().
waitall() is now just a compatibility alias for that variation. This makes
wait() consistent with wait_each(), wait_each_success() and
wait_each_exception().

Add tests for this new functionality.