greendns: Return answers from /etc/hosts despite nameserver errors
When calling getaddrinfo with an unspecified address family, we:

 * try to get an IPv6 address for the host, first trying /etc/hosts and
   failing that, using any external nameserver, then
 * try to get an IPv4 address for it, with the same /etc/hosts ->
   external nameserver priority.

Note that if the host can be resolved using /etc/hosts but only to one
of the two address families, we still attempt to contact the external
nameserver to resolve it in the other family.

Previously, if this timed out or raised a DNSException other than
NXDOMAIN, EAI_EAGAIN_ERROR or EAI_NODATA_ERROR would be raised even
though we could resolve the host. Now, those errors will only be raised
if neither attempt at resolution succeeded.