Make hubs use a monotonic clock by default
Use a monotonic clock by default, so that hubs are not affected by
changes of system time (otherwise events will appear to stop, if
time is adjusted backwards).

The suggestion here is to add a new dependency on monotonic, which
is needed to make this work on Python versions prior to 3.3. This
will allow eventlet to do "the right thing" by default, so that
consuming projects do not have to do that on their own (e.g.
oslo_service - https://bugs.launchpad.net/oslo.service/+bug/1510234).

Fixes #132