Fix parsing of Windows hosts file.
The default Windows hosts file starts something like this:

    # Copyright (c) 1993-2009 Microsoft Corp.
    #
    # This is a sample HOSTS file used by Microsoft TCP/IP for Windows.
    ...

But `greendns.HostsResolver.LINES_RE` doesn't admit the possibility of a
comment starting at the beginning of a line. It produces a sequence consisting
of data from the comment lines as well as (eventually) the real content of the
hosts file.

This has been ignored since `is_ipv4_addr()` and `is_ipv6_addr()` catch
`dns.exception.SyntaxError` and return `False`. But in a runtime environment
that encounters #413, `dns.exception.SyntaxError` is not caught, and the
import fails.

Changing '+' to '*' allows `HostsResolver._readlines()` to recognize and skip
comment lines. However, that produces empty strings in the sequence. Filter
out the empty strings by passing the sequence through
`itertools.ifilter(None)`.