Use external dependencies for six, monotonic, dnspython.
Remove the snapshots of these packages embedded in eventlet/support; instead
use setup.py's install_requires and dependency_links options to fetch them
into the top-level package namespace.

Because embedded snapshots are inherently version-locked, these dependencies
are likewise version-locked.

As of 2018-05-01, dnspython version 1.16.0 that was embedded in
eventlet/support is not yet available from PyPI. bin/pull-dnspython referenced
a specific commit in the upstream github.com repository; the same commit ID in
the same repository is referenced in the dependency_links URL. However, it
seems that this mechanism requires the user to specify pip's
--process-dependency-links switch.