wsgi: Fix header capitalization on py3
str.capitalize() and str.upper() respect unicode capitalization rules on
py3, while py2 just translates a-z to A-Z.

At best, this may cause confusion and unexpected behaviors, such as
when '\xdf' (a Latin1-encoded ß) becomes 'SS'; at worst, this causes
UnicodeEncodeErrors and the server fails to reply, such as when '\xff'
(a Latin1-encoded ÿ) becomes '\u0178' which does not map back into
Latin1.

Now, convert everything to bytes before capitalizing so just a-z and A-Z
are affected on both py2 and py3.