Issue #526: Introduce a self-serving HTTPS test.
The new monkey_patch_socket_https_526.py test script depends on an 'openssl'
executable in the PATH. Using this, it creates temporary certificate files
with which to populate an SSLContext so that a local server can accept HTTPS
connections.

We run it by tests.openssl_test.test_https() calling run_isolated() to launch
a new Python interpreter without eventlet imports. For the same reason, the
test script uses multiprocessing to launch the server as a separate process.
Only the client code imports eventlet and uses monkey_patch(socket=True),
which drives the reported issue.