Fix leak of epoll file descriptors
When a python process creates a new thread that uses an eventlet hub
(e.g. when using `trampoline`) a new epoll file descriptor is allocated.
If the thread now stops after a short amount of work the epoll file
descriptor is not closed causing a leak.

If the python process (but not the thread) is long living and regularly
spawns a lot of short-lived threads then this can cause an exhaustion of
file descriptors.

To solve this the epoll file descriptor is now no longer allocated on
hub creation but only when a listener is actually added. We also now
close the file descriptor if there are no listeners left.