Handle transport endpoint shutdown in conditions
ESHUTDOWN errors types are raised when a socket
cannot send after transport endpoint is shutdown [1].

BrokenPipeError correspond to errno EPIPE and ESHUTDOWN.
BrokenPipeError are raised when trying to write on a pipe
while the other end has been closed, or trying to write
on a socket which has been shutdown for writing [2].

If we specificaly handle EPIPE [3] in our exception management
we should also handle ESHUTDOWN in the same way. A BrokenPipeError.

This patch made the addition of the ESHUTDOWN each time EPIPE
is specificaly handled.

[1] https://docs.python.org/fr/3/library/errno.html#errno.ESHUTDOWN
[2] https://docs.python.org/fr/3/library/exceptions.html#BrokenPipeError
[3] https://docs.python.org/fr/3/library/errno.html#errno.EPIPE