Implement file descriptor removal and waiting in asyncio hub
We recently observed (#912) difference in the behaviors between
the asyncio hub and the other eventlet hubs.

The asyncio hub behaved differenly because bad file descriptor
were not properly removed from the list of readers and writers
listeners.

To clean these bad file descriptors we have to implement
an awaitable wait method specific to the asyncio hub

This patch should fix #912.

I prefer to propose this patch as a separated patch
to isolate topics and simplify reviews.

https://github.com/eventlet/eventlet/pull/912