Refactor congruence checks based on assert at runtime
Assert statements are completely eliminated when the
python interpreter is ran with the optimization flags [1].

Those checks are not an option and should be executed
not matter the execution context.

This patch refactor those checks to not rely anymore
on the assert statement.

This patch is mainly focused on hub checks. Other
assert statements are still present, I'll refactor
them later.

Refactoring this kind of code ensure to not waste
time managing issues and debugging bugs related to
the optimization features of the interpreter.

[1] https://docs.python.org/3/using/cmdline.html#cmdoption-O